Rename `src.gui` -> `src.qt`.
