+++ .gitignore
@@ -1,5 +1,6 @@
 # Files generated by the tool.
 ./config.json
+./credentials.json
 
 # Byte-compiled / optimized / DLL files
 __pycache__/
