+++ src/qt/functions.py
